+++ server/src/signatureHelp.ts
@@ -14,15 +14,22 @@
  * limitations under the License.
  */
 
-import { SignatureHelp, SignatureHelpParams } from "vscode-languageserver";
-import { nodeFinder } from "./helpers";
+import {
+  ParameterInformation,
+  SignatureHelp,
+  SignatureHelpParams,
+  SignatureInformation,
+} from "vscode-languageserver";
+import { getCMacroCall, nodeFinder } from "./helpers";
 import { ContextAware } from "./runtimeEvaluator";
 import { SearchableResult } from "./types";
 import { Property } from "./context/property";
+import { ASTBase } from "./ast/base";
+import { FunctionDefinition } from "./ast/cPreprocessors/functionDefinition";
+import { CIdentifier } from "./ast/cPreprocessors/cIdentifier";
 
 function getPropertySignatureHelp(
-  result: SearchableResult | undefined,
-  context: ContextAware
+  result: SearchableResult | undefined
 ): SignatureHelp | undefined {
   if (!result || !(result.item instanceof Property)) {
     return;
@@ -38,6 +45,63 @@ function getPropertySignatureHelp(
   return signatureHelp;
 }
 
+async function getMacroFuncSignatureHelp(
+  result: SearchableResult | undefined
+): Promise<SignatureHelp | undefined> {
+  const macroCall = getCMacroCall(result?.ast);
+  if (
+    !macroCall ||
+    !macroCall.params.length ||
+    !result?.ast ||
+    result.ast === macroCall.functionName
+  ) {
+    return;
+  }
+
+  const lastParser = (await result.runtime.context.getAllParsers()).at(-1)!;
+  const macroDef = lastParser.cPreprocessorParser.macros.get(
+    macroCall.functionName.name
+  )?.macro.identifier;
+
+  if (!macroDef || !(macroDef instanceof FunctionDefinition)) {
+    return;
+  }
+
+  const signatureArgs: ParameterInformation[] = macroDef.params.map((p) => ({
+    label: p instanceof CIdentifier ? p.name : "...",
+  }));
+
+  let param = macroCall.params.find(
+    (p) => p && (result.ast === p || p.isAncestorOf(result.ast))
+  );
+
+  param ??= result.afterAst
+    ? macroCall.params.find(
+        (p) => p && (result.afterAst === p || p.isAncestorOf(result.afterAst!))
+      )
+    : undefined;
+  param ??= result.beforeAst
+    ? macroCall.params.find(
+        (p) =>
+          p && (result.beforeAst === p || p.isAncestorOf(result.beforeAst!))
+      )
+    : undefined;
+
+  return {
+    signatures: [
+      SignatureInformation.create(
+        `${macroCall.functionName.name}(${signatureArgs
+          .map((arg) => arg.label)
+          .join(", ")})`,
+        undefined,
+        ...signatureArgs.flat()
+      ),
+    ],
+    activeSignature: 0,
+    activeParameter: param ? macroCall.params.indexOf(param) : undefined,
+  };
+}
+
 export async function getSignatureHelp(
   location: SignatureHelpParams,
   context: ContextAware | undefined
@@ -46,7 +110,8 @@ export async function getSignatureHelp(
 
   return (
     await nodeFinder(location, context, async (locationMeta) => [
-      getPropertySignatureHelp(locationMeta, context),
+      (await getMacroFuncSignatureHelp(locationMeta)) ||
+        getPropertySignatureHelp(locationMeta),
     ])
   ).at(0);
 }